chore: rename agent loader files
