+++ packages/core/src/loader/agent-js.ts
