+++ packages/core/src/loader/agent-yaml.ts
